Docs Update setup.mdx (types of WeaveDB SDKs)
Hello WeaveDB team,

As a Weave Ambassador, I've noticed a mismatch in the documentation regarding the number of SDK types supported by WeaveDB. The current documentation states that WeaveDB has "4 types of SDKs," but then only provides details for two SDKs. To avoid any confusion for our users, we need to clarify the information presented in the documentation.

In the current documentation, the following text is provided:

"Set up SDK

WeaveDB has 4 types of SDKs.

- a full WeaveDB SDK wrapping the Warp SDK for the web weavedb-sdk
- a full SDK for NodeJS weavedb-sdk-node
The query APIs for both the SDKs are the same."

However, only two SDKs are detailed in the documentation, namely, weavedb-sdk and weavedb-sdk-node. There seems to be a mismatch between the statement "WeaveDB has 4 types of SDKs" and the subsequent listing of only two SDKs.

Solution:

To address this issue, I have updated the documentation to accurately reflect the available 2 SDK types. If there are indeed two additional SDKs beyond weavedb-sdk and weavedb-sdk-node, it would be good to include information on them as well.